+++ EkahauRevitPlugin/VersionInfo.cs
@@ -7,7 +7,7 @@ namespace EkahauRevitPlugin
     /// </summary>
     public static class VersionInfo
     {
-        public const string Version     = "2.5.15";
+        public const string Version     = "2.5.16";
         public const string ReleaseDate = "2026-05-04";
         public const string RepoUrl     = "https://github.com/nicoliu0212/Ekahau-Revit-Transfer";
         public const string LicenseUrl  = "https://github.com/nicoliu0212/Ekahau-Revit-Transfer/blob/main/LICENSE";